Move App.js to src directory
